Add utils module to handle .ini and files in general. Add metadata module rather than having it embedded in a generic module
